BUG: Refs #212.  Work towards exporting batchmake scripts.
This commit breaks out much functionality into KWBatchmakeComponent.php and
also further into a library (KWUtils.php).  Included are tests for both of
these, along with testing data.  The functionality provided would allow
for creating a tmp work directory in the batchmake module, copying in
the needed .bms and .bmm files, compiling the batchmake script, creating
the condor DAG and scripts from the .bms file, and submitting the DAG to
condor.