BUG: Refs #212.  Fixed bug in chmod warning, and rewrite of tests.
There was a bug when a file was tested to see if it could be chmod-ed,
a warning was issued if the file could not be chmoded and this broke
json parsing.  Now I've changed this to first check the UID of the process,
and see if that is the same as the UID of the file, if not, then don't perform
the chmod test and hence don't generate the warning.

Also in this commit is the creation of the BatchmakeControllerTest superclass
and changes in the batchmake tests to use this superclass.  This superclass
helps streamline batchmake testing, so the user doesn't have to set as many
properties manually to test the batchmake module.