ENH: refs #0322. Allow batchmake tests to pass test configs.
Batchmake tests and KWBatchmakeComponent had assumed an install was performed,
but this isn't the case for a dashboard machine.  Now the controller tests
will set a test config through the Zend_Registry, and the KWBatchmakeComponent
will check if it is in testing mode if it cannot find an installed config.